move label shelf functions to a separate file
