+++ pyproject.toml
@@ -17,7 +17,7 @@ line-length = 100
 
 [tool.ruff.lint]
 select = ["ALL"]
-ignore = ["EM101", "EM102", "RET504", "TRY003"]
+ignore = ["EM101", "EM102", "RET504", "TRY003", "S101"]
 
 [tool.ruff.lint.per-file-ignores]
 # Ignore folowing rules for tests:
